tc_build: llvm: Ensure LLVM_LINK_LLVM_DYLIB remains off
Currently, this is disabled by default but LLVM upstream is talking
about flipping this value to ON for certain platform. It hurts the speed
of the compiler, which is one of tc-build's primary focuses with the
easy integration of PGO, BOLT, and LTO. Keep this option off to maintain
the status quo.

Link: https://discourse.llvm.org/t/rfc-llvm-link-llvm-dylib-should-default-to-on-on-posix-platforms/85908
Signed-off-by: Nathan Chancellor <nathan@kernel.org>